+++ hu.bme.mit.incqueryd.core/src/main/java/hu/bme/mit/incqueryd/rete/comparison/ConditionExpression.java
@@ -11,16 +11,16 @@ public class ConditionExpression {
 
     protected int index;
     protected ComparisonOperator comparisonOperator;
-    protected int operand;
+    protected long operand;
 
-    public ConditionExpression(final int index, final ComparisonOperator comparisonOperator, final int operand) {
+    public ConditionExpression(final int index, final ComparisonOperator comparisonOperator, final long operand) {
         this.index = index;
         this.comparisonOperator = comparisonOperator;
         this.operand = operand;
     }
 
     public boolean satisfiesCondition(final Tuple tuple) {
-        final Integer tupleField = (Integer) tuple.get(index);
+        final Long tupleField = (Long) tuple.get(index);
         return Compare.compare(tupleField, operand, comparisonOperator);
     }
